Use a map join approach to accepted bet confirm
Helps avoid a (impossible/unlikely) case where a saltbot "bet accept"
message is processed before the actual bet message is processed.

Also, fixes a bug where the first bet of round wouldn't update properly
because the bet attempt would be recorded, then everything would be
reset by the logic in the saltbot "accept" processing for that bet, thus
voiding it.